Implement authorization checks (very basic, not ideal) for updating and deleting ideas based on username

+++ routes/ideas.js
@@ -51,18 +51,28 @@ router.post("/", async (req, res) => {
 // Update Idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
-        },
-      },
-      { new: true }
-    );
+    const idea = await Idea.findById(req.params.id);
 
-    res.json({ success: true, data: updatedIdea });
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
+        },
+        { new: true }
+      );
+      res.json({ success: true, data: updatedIdea });
+    } else {
+      res.status(403).json({
+        // Usernames does not match
+        success: false,
+        error: "You are not authorized to update this resource",
+      });
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "Something went wrong" });
@@ -72,8 +82,19 @@ router.put("/:id", async (req, res) => {
 // Delete Idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.send({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body?.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      res.send({ success: true, data: {} });
+    } else {
+      // Usernames does not match
+      res.status(403).json({
+        success: false,
+        error: "You are not authorized to delete this resource",
+      });
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "Something went wrong" });
